requirements.txt
feedparser>=6.0.8
openai>=0.27.0
python-dotenv>=0.19.2
requests>=2.28.2